Fix crontab
Apparently `/etc/crontab` is special in that it requires specifying which user is to run each command.
Because it was misconfigured for so long, the server has seemingly run without a functional `crond` for a long time.  I can only hope that the cronjobs *not* specified in `/etc/crontab` worked fine throughout this period.

Also worth noting: Because the root user was disabled, it seems that it was impossible to send mail to it. Mail for these cronjobs is now being routed to the minecraft user, since they own 99% of the jobs.