Reworked ranks to be compliant with the new EULA
The new EULA, which we never agreed to and which was not out at the time that our version of Minecraft was released, stipulates that donors cannot receive any material advantage over non-donors.  Accordingly, the entire pre-existing concept of the donor rank had to be scrapped.  The result of this is that most cosmetic features have been moved into the donor rank, the donor rank's old abilities have been moved into other ranks, and job leaving/joining has been made available to everyone, though at the cost of a 10% loss in skill if not maxxed.